Rename quda_cuda_api.* to quda_api.h and move CUDA implementation to targets/cuda directory
