+++ lib/CMakeLists.txt
@@ -95,7 +95,7 @@ set (QUDA_OBJS
   gauge_fix_ovr_extra.cu gauge_fix_fft.cu gauge_fix_ovr.cu
   pgauge_det_trace.cu clover_outer_product.cu
   clover_sigma_outer_product.cu momentum.cu gauge_qcharge.cu
-  quda_cuda_api.cpp deflation.cpp checksum.cu
+  deflation.cpp checksum.cu
   instantiate.cpp version.cpp )
 # cmake-format: on
